backend/requirements.txt
Flask
Flask-Session
Flask-CORS
python-dotenv
bcrypt
mysql-connector-python 
didkit
